refactor: moved GameObject to controller folder
