+++ src/main/java/io/github/agentseek/controller/GameObject.kt
@@ -1,7 +1,8 @@
-package io.github.agentseek.model
+package io.github.agentseek.controller
 
 
 import io.github.agentseek.common.Point2d
+import io.github.agentseek.model.World
 import io.github.agentseek.model.components.Component
 import io.github.agentseek.model.physics.HitBox
