+++ src/main/java/io/github/agentseek/model/components/CollisionWithPlayerComponent.kt
@@ -1,8 +1,8 @@
 package io.github.agentseek.model.components
 
 import io.github.agentseek.controller.events.ItemPickUpEvent
-import io.github.agentseek.model.GameObject
-import io.github.agentseek.model.Type
+import io.github.agentseek.controller.GameObject
+import io.github.agentseek.controller.Type
 import io.github.agentseek.model.World
 import ryleh.controller.events.EnemyCollisionEvent
