+++ src/main/java/io/github/agentseek/model/components/ShooterComponent.kt
@@ -3,7 +3,7 @@ package io.github.agentseek.model.components
 import io.github.agentseek.common.Point2d
 import io.github.agentseek.common.Vector2d
 import io.github.agentseek.controller.Entity
-import io.github.agentseek.model.GameObject
+import io.github.agentseek.controller.GameObject
 import io.github.agentseek.model.World
 
 /**
