+++ src/test/java/test/java/GameOverTest.java
@@ -9,7 +9,7 @@ import ryleh.common.Circle2d;
 import ryleh.controller.events.Event;
 import ryleh.controller.events.EventListener;
 import io.github.agentseek.controller.events.GameOverEvent;
-import io.github.agentseek.model.GameObject;
+import io.github.agentseek.controller.GameObject;
 import ryleh.model.GameObjectImpl;
 import ryleh.model.Type;
 import ryleh.model.World;
